refactor(src create-json-from-handlebars-templates create-json-from-templates-and-zone-data): split
BREAKING CHANGE: splitting the files involved changing the exported function name from
createJSONFromHandlebarsTemplatesAndZoneData to crateJSONFromTemplatesAndZoneData - no mention of
handlebars templates. The plan is to support another type of template. The timezones folder has also
been update to only include the latest templates, as the older timezones were incorrect, because of
the country parsing bug.
