+++ src/create-json-from-handlebars-templates/create-json-from-handlebars-template.spec.ts
@@ -1,105 +1,12 @@
 import {
     createJSONFromHandlebarsTemplates,
-    createJSONFromHandlebarsTemplatesAndZoneData,
     ICreateJSONFromHandlebarsTemplatesParams
 } from './create-json-from-handlebars-templates';
 import * as path from 'path';
 import {readFileAsync} from '../util/util';
-import {IANATzDataFiles} from '../get-iana-tz-data/get-iana-tz-data';
 import {sampleExtractedData} from './test-extracted-data';
 
 describe('create-json-from-handlebars-template', () => {
-    describe('.createJSONFromHandlebarsTemplatesAndZoneData', () => {
-        it('should call .createJSONFromHandlebarsTemplates with default parameters and get files and zone data', async() => {
-            const createJSONFromHandlebarsTemplatesSpy = jasmine
-                .createSpy('createJSONFromHandlebarsTemplates');
-
-            const getIANATzDataStub = jasmine.createSpy('getIANATzDataSpy')
-                .and.returnValue(fakeIANAData());
-
-            await createJSONFromHandlebarsTemplatesAndZoneData(
-                {} as any, // will use default settings
-                getIANATzDataStub,
-                createJSONFromHandlebarsTemplatesSpy,
-            );
-
-            // check default params
-            expect(createJSONFromHandlebarsTemplatesSpy).toHaveBeenCalledTimes(1);
-            const firstCallParams: ICreateJSONFromHandlebarsTemplatesParams =
-                createJSONFromHandlebarsTemplatesSpy.calls.argsFor(0)[0];
-
-            expect(firstCallParams.handlebarsTemplateFileNames).toContain('all-fields/all-fields-v2.hbs');
-            expect(firstCallParams.extractedZoneData.zones[0].countryCodes[0]).toEqual('AD');
-            expect(firstCallParams.templatesPath).toContain('templates');
-            expect(firstCallParams.zoneFileName).toContain('zone1970.tab');
-            expect(firstCallParams.saveDirectory).toContain('timezones');
-        });
-
-        it('should call .createJSONFromHandlebarsTemplates with custom parameters and get files and zone data', async() => {
-            const createJSONFromHandlebarsTemplatesSpy = jasmine
-                .createSpy('createJSONFromHandlebarsTemplates');
-
-            const getIANATzDataStub = jasmine.createSpy('getIANATzDataSpy');
-
-            const extractTzDataSpy = jasmine.createSpy('extractTzData')
-                .and.returnValue(new Promise((resolve, reject) => {
-                    resolve({foo: 'bar'})
-                }));
-
-            const walkAsyncSpy = jasmine
-                .createSpy('walkAsyncSpy')
-                .and.returnValue(new Promise((resolve, reject) => { resolve(['foo.hbs']) }));
-
-            await createJSONFromHandlebarsTemplatesAndZoneData(
-                {
-                    templatesPath: 'fooTemplates',
-                    saveDirectory: 'fooSaveDirectory',
-                    zoneFileNames: ['foo.bar']
-                } as any, // settings
-                getIANATzDataStub,
-                createJSONFromHandlebarsTemplatesSpy,
-                extractTzDataSpy,
-                walkAsyncSpy
-            );
-
-            expect(getIANATzDataStub).toHaveBeenCalledWith({filesToExtract: ['foo.bar']});
-
-            // check default params
-            expect(createJSONFromHandlebarsTemplatesSpy).toHaveBeenCalledTimes(1);
-            const firstCallParams: ICreateJSONFromHandlebarsTemplatesParams =
-                createJSONFromHandlebarsTemplatesSpy.calls.argsFor(0)[0];
-
-            expect(firstCallParams.handlebarsTemplateFileNames).toEqual(['foo.hbs']);
-            expect(firstCallParams.extractedZoneData).toEqual({foo: 'bar'} as any);
-            expect(firstCallParams.templatesPath).toEqual('fooTemplates');
-            expect(firstCallParams.zoneFileName).toEqual('foo.bar');
-            expect(firstCallParams.saveDirectory).toEqual('fooSaveDirectory');
-        });
-
-        it('should filter the templates path to only use .hbs (handlebars) files', async() => {
-            const createJSONFromHandlebarsTemplatesSpy = jasmine.createSpy('createJSONFromHandlebarsTemplates');
-            const getIANATzDataStub = jasmine.createSpy('getIANATzDataSpy');
-            const extractTzDataSpy = jasmine.createSpy('extractTzData');
-
-            const walkAsyncSpy = jasmine
-                .createSpy('walkAsyncSpy')
-                .and.returnValue(new Promise((resolve, reject) => { resolve(['foo.hbs', 'bar.json', 'baz.hbs']) }));
-
-            await createJSONFromHandlebarsTemplatesAndZoneData(
-                {} as any, // settings
-                getIANATzDataStub,
-                createJSONFromHandlebarsTemplatesSpy,
-                extractTzDataSpy,
-                walkAsyncSpy
-            );
-
-            const firstCallParams: ICreateJSONFromHandlebarsTemplatesParams =
-                createJSONFromHandlebarsTemplatesSpy.calls.argsFor(0)[0];
-
-            expect(firstCallParams.handlebarsTemplateFileNames).toEqual(['foo.hbs', 'baz.hbs']);
-        })
-    });
-
     describe('.createJSONFromHandlebarsTemplates', () => {
         const params: ICreateJSONFromHandlebarsTemplatesParams = {
             handlebarsTemplateFileNames: ['aaa.hbs'],
@@ -274,19 +181,8 @@ describe('create-json-from-handlebars-template', () => {
 
             expect(error!.message.toString()).toContain('Could not write error file:');
             expect(error!.message.toString()).toContain('Trying to write to path: saveDirFoo/error.txt');
-
         })
     })
 });
 
-async function fakeIANAData(): Promise<IANATzDataFiles> {
-    const zone1970path = path.join(__dirname, 'test-zone1970.tab');
-    const zone1970tab = await readFileAsync(zone1970path);
-    const zonePath = path.join(__dirname, 'test-zone.tab');
-
-    return {
-        version: 'TEST',
-        ['zone1970.tab']: zone1970tab.toString(),
-    }
-}
