+++ src/index.ts
@@ -1,3 +1,3 @@
 export {getIANATzData} from './get-iana-tz-data/get-iana-tz-data';
 export {loadIANATzJsonFile} from './load-iana-tz-json-file/load-iana-tz-json-file';
-export {createJSONFromHandlebarsTemplatesAndZoneData} from './create-json-from-handlebars-templates/create-json-from-handlebars-templates'
+export {createJSONFromTemplatesAndZoneData} from './create-json-from-templates-and-zone-data/create-json-from-templates-and-zone-data';
